Get details about non-CRAN package installation

Get installation code of packages that either should be installed not from CRAN or a modified code should be installed.

get_pkgs_non_cran_installation_details(
  use_local_list = getOption("bio.use_local_list", FALSE)
)

Arguments

use_local_list
(logical) If TRUE, the list, which is locally installed in the folder of package bio ("local list"), is used. If FALSE, the list on "GitHub" repository of the package is used. It is recommended using the online version of the list, as it may contain more recent changes. Optiom bight be set globally by, e.g., options(bio.use_local_list = TRUE).

Value

Dataframe with columns:

"package" package name,

"install_from" source to install from. Currently supported values are "code" and "github"

"details" either installation code of repository name

"notes" notes for user.

Details

If the file has column ignore with value TRUE, the line is not included in the output.

See also

Examples

# NOTE: It is not recommended to use the local lists as they might be out of date.
options(bio.use_local_list = TRUE)

head(get_pkgs_non_cran_installation_details())

#>        package install_from               details notes
#> 1  addin.tools       github   GegznaV/addin.tools  <NA>
#> 2   addins.rmd       github    GegznaV/addins.rmd  <NA>
#> 3    addins.rs       github     GegznaV/addins.rs  <NA>
#> 4      adventr       github profandyfield/adventr  <NA>
#> 5 backup.tools       github  GegznaV/backup.tools  <NA>
#> 6          bio       github           mokymai/bio  <NA>