Get required version of packages.

Get required version of packages from a list in a file.

get_pkgs_req_version(use_local_list = getOption("bio.use_local_list", FALSE))

Arguments

use_local_list
(logical) If TRUE, the list, which is locally installed in the folder of package bio ("local list"), is used. If FALSE, the list on "GitHub" repository of the package is used. It is recommended using the online version of the list, as it may contain more recent changes. Optiom bight be set globally by, e.g., options(bio.use_local_list = TRUE).

Value

Dataframe with columns "package" and "required_version".

See also

Examples

# NOTE: It is not recommended to use the local lists as they might be out of date.
options(bio.use_local_list = TRUE)

head(get_pkgs_req_version())

#>        package required_version notes_version
#> 1  addin.tools            0.0.8          <NA>
#> 2   addins.rmd           0.0.10          <NA>
#> 3    addins.rs            0.0.7          <NA>
#> 4 backup.tools            0.0.1          <NA>
#> 5          bio       0.0.0.9024          <NA>
#> 6     bookdown             0.17          <NA>