Reset RStudio state in GMC R209 and clear environment

This function:

Resets RStudio state and user preferences (incl. color scheme)

Clears function history, plot history, console, recent project list, etc.

Closes unnecessary windows

Resets custom keybindings to "bio-default"

Resets R Markdown and R snippets to defaults in package "snippets"

Creates folder "~/R/Darbinis" and starts using it as working directory when no project is used.

Clears (if possible)/creates folder "BS-2020" on Desktop.

rstudio_reset_gmc(..., force_update_dictionaries = FALSE)

Arguments

...
Further arguments for advanced users only.

force_update_dictionaries
(logical) If TRUE, the dictionaries are forced to be downloaded/updated.

Details

The function works only in GMC R209, GMC C255 and ITPC EC-108 classrooms.

Examples