Manage RStudio projects

read_projects() - reads file with projects and list their names and paths.

get_path_recent_proj_list() -- gets path to the file with the list of recent RStudio projects.

get_path_personal_proj_list() -- gets path to the file with the list of personal RStudio projects.

get_projs_recent() -- lists recent RStudio projects.

get_projs_personal() -- lists personal RStudio projects.

get_projs_all() -- lists both recent and personal RStudio projects.

get_proj_names() -- lists RStudio projects names (as a character vector)

read_projects(file, sort_by = FALSE)

get_path_recent_proj_list()

get_path_personal_proj_list(create = FALSE)

get_projs_recent(sort_by = FALSE)

get_projs_personal(sort_by = FALSE)

get_projs_all()

get_proj_names(file = get_path_recent_proj_list(), sort_by = FALSE)

Arguments

file
(character) Path to file with RStudio projects.

sort_by
("name"|"path"|FALSE)

create
(logical) If TRUE and file does not exist, the file is created.

Examples