Prepare Files For Submission (DataCamp Task)

The function creates a folder (e.g.,U00-atsakymai-ikelimui) and copies files indicated in sertifikatas and konspektas to that folder. The copied files are properly renamed. If needed, the function opens the folder and assignment submission webpage.

u_prepare_assignment_dc(
  uzduoties_nr,
  varianto_nr,
  specialybe,
  pavarde,
  vardas,
  sertifikatas,
  konspektas,
  open_dir = FALSE,
  emokymai_id = NULL
)

Arguments

uzduoties_nr

(integer)

varianto_nr

(integer)

specialybe

(string)

pavarde

(string)

vardas

(string)

sertifikatas

(string) Path to PDF file (certificate issued by DataCamp).

konspektas

(string) Path to PDF file.

open_dir

(TRUE|FALSE) Open the directory with submission files.

emokymai_id

(integer) Open the directory with submission files.

Examples

if (FALSE) {
  u_prepare_assignment_dc(
    uzduoties_nr = 1,
    varianto_nr  = 000,
    specialybe   = "molekuline biologija",
    pavarde      = "Pavarde",
    vardas       = "Vardas Antras",
    sertifikatas = "pratybos/certificate.pdf",
    konspektas   = "pratybos/konspektas.pdf"
  )
}