Create Filename for Type U Task

Create a name for a file or folder which should be submitted as a "U" task.

u_create_filename(
  uzduoties_nr,
  varianto_nr,
  specialybe,
  pavarde,
  vardas,
  patikslinimas = "",
  dokumento_formatas = NULL
)

u_parse_filename(x)

u_check_filename(x)

Arguments

uzduoties_nr

(integer) The number of U task.

varianto_nr

(integer) The variant of the task. Each student has a unique personal variant number.

specialybe

(character) Name of study programme.

pavarde

(character) Family name (names).

vardas

(character) Sure name (names).

patikslinimas

(character) Additional information.

dokumento_formatas

(character) File name extension.

x

(character) A file name to parse or to check if it's structure is correct.

Value

u_create_filename() returns sting with formatted file name.

u_parse_filename() returns data frame with details extracted from the filename.

u_check_filename() returns TRUE if the structure of file name is correct and FALSE otherwise.

Examples

u_create_filename(
  uzduoties_nr       = 1,
  varianto_nr        = 000,
  specialybe         = "molekuline biologija",
  pavarde            = "Pavarde",
  vardas             = "Vardas Antras",
  patikslinimas      = "sertifikatas",
  dokumento_formatas = "pdf"
)
#> U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf

u_create_filename(
  uzduoties_nr       = 10,
  varianto_nr        = 000,
  specialybe         = "molekuline biologija",
  pavarde            = "Pavarde",
  vardas             = "Vardas Antras"
)
#> U10-v000-[molbio]-[Pavarde]-[Vardas-Antras]

x <- "U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf"

u_check_filename("U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf")
#> [1] TRUE
u_parse_filename("U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf")
#>             Kategorija              Reiksme
#> 1       uzduoties_nr:                   U01
#> 2        varianto_nr:                   000
#> 3         specialybe:  molekuline biologija
#> 4            pavarde:               Pavarde
#> 5             vardas:         Vardas Antras
#> 6      patikslinimas:          sertifikatas
#> 7 dokumento_formatas:                   pdf

u_check_filename("U03-v000-[molekuline]-[Pavarde]-[Vardas].zip")
#> [1] TRUE
u_parse_filename("U03-v000-[molekuline]-[Pavarde]-[Vardas].zip")
#>             Kategorija              Reiksme
#> 1       uzduoties_nr:                   U03
#> 2        varianto_nr:                   000
#> 3         specialybe:  molekuline biologija
#> 4            pavarde:               Pavarde
#> 5             vardas:                Vardas
#> 6      patikslinimas:                      
#> 7 dokumento_formatas:                   zip

u_check_filename("U03-v000-[biofiz]-[Pavarde]-[Vardas]")
#> [1] TRUE
u_parse_filename("U03-v000-[biofiz]-[Pavarde]-[Vardas]")
#>             Kategorija    Reiksme
#> 1       uzduoties_nr:         U03
#> 2        varianto_nr:         000
#> 3         specialybe:   biofizika
#> 4            pavarde:     Pavarde
#> 5             vardas:      Vardas
#> 6      patikslinimas:            
#> 7 dokumento_formatas:  { FOLDER }

# Incorrect family name:
u_parse_filename("U03-v000-[molekuline]-[Pavarde3]-[Vardas].zip")
#>             Kategorija                          Reiksme
#> 1       uzduoties_nr:                              <NA>
#> 2        varianto_nr:                              <NA>
#> 3         specialybe:     (?) UNIDENTIFIED VALUE { NA }
#> 4            pavarde:                              <NA>
#> 5             vardas:                              <NA>
#> 6      patikslinimas:     (?) UNIDENTIFIED VALUE { NA }
#> 7 dokumento_formatas:  (?) UNSUPPORTED EXTENSION { NA }